+++ packages/react-compat-moveable/rollup.config.js
@@ -15,6 +15,8 @@ const external = {
     "@moveable/matrix": "@moveable/matrix",
     "@scena/dragscroll": "@scena/dragscroll",
     "css-to-mat": "css-to-mat",
+    "overlap-area": "overlap-area",
+    "@scena/matrix": "@scena/matrix",
 };